refactor: create smaller components
