+++ src/assets/css/base.css
@@ -88,6 +88,8 @@
   @apply border-transparent;
 }
 
+html { @apply scroll-smooth }
+
 body {
   @apply font-sans;
   @apply relative;
