+++ src/components/ExperienceDetails.vue
@@ -1,0 +1,21 @@
+<script setup lang="ts">
+import Markdown from "vue3-markdown-it";
+import type { ToolData } from '@/helpers/experience';
+
+defineProps<{
+  markdown: string;
+  tools: ToolData[];
+}>();
+</script>
+
+<template>
+  <div ref="tblx-exp" class="experience">
+    <Markdown class="mb-4 mt-2 flex flex-col gap-4 " :source="markdown" :html="true" />
+
+    <div class="grid-cols-3 sm:grid-cols-4 gap-2 text-sm content-center w-full grid md:hidden">
+      <div v-for="tool in tools" :key="tool.name" :class="`tool tool-${tool.type} active`">
+        {{ tool.name }} <i :class="tool.icon"></i>
+      </div>
+    </div>
+  </div>
+</template>
